+++ com.unity.netcode.gameobjects/Tests/Editor/Serialization/FastBufferReaderTests.cs
@@ -1579,5 +1579,21 @@ namespace Unity.Netcode.EditorTests
                 Assert.AreEqual(reader.Handle->AllowedReadMark, 25);
             }
         }
+
+        [Test]
+        public unsafe void WhenUsingArraySegment_ConstructorHonorsArraySegmentConfiguration()
+        {
+            var bytes = new byte[] { 0, 1, 2, 3 };
+            var segment = new ArraySegment<byte>(bytes, 1, 3);
+            var reader = new FastBufferReader(segment, Allocator.Temp);
+
+            var readerArray = reader.ToArray();
+            Assert.True(readerArray.Length == bytes.Length - 1, $"Array of reader should have a length of {bytes.Length - 1} but was {readerArray.Length}!");
+            for (int i = 0; i < readerArray.Length; i++)
+            {
+                Assert.True(bytes[i + 1] == readerArray[i], $"Value of {nameof(readerArray)} at index {i} is {readerArray[i]} but should be {bytes[i + 1]}!");
+            }
+            reader.Dispose();
+        }
     }
 }