+++ com.unity.netcode.gameobjects/Tests/Editor/Transports/BatchedSendQueueTests.cs
@@ -2,7 +2,9 @@ using System;
 using NUnit.Framework;
 using Unity.Collections;
 using Unity.Netcode.Transports.UTP;
+#if !UTP_TRANSPORT_2_0_ABOVE
 using Unity.Networking.Transport;
+#endif
 
 namespace Unity.Netcode.EditorTests
 {